refactor: removes HOC, adds shortcut to keyMap
- Removed the HOC file and its association from the timeline files
- Refactored the scrollTo logic to find the timeline's first article
- Adds refreshFeed to the ui handler and keyMap
